requirements/base.txt
coreapi==2.3.0
Django==1.11.4
django-cors-headers==2.0.2
djangorestframework==3.6.3
Markdown==2.6.8
Pillow==4.1.1
Pygments==2.2.0
